# app/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This swagger describes every endpoint of this project.",
        "title": "Augustin Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "GNU Affero General Public License",
            "url": "https://www.gnu.org/licenses/agpl-3.0.txt"
        },
        "version": "0.0.1"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/api/vendors/comments/{id}/": {
            "delete": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Delete vendor comment",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete vendor comment",
                "operationId": "deleteVendorComment",
                "responses": {}
            }
        },
        "/api/vendors/locations/{id}/": {
            "delete": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Delete vendor location",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete vendor location",
                "operationId": "deleteVendorLocation",
                "responses": {}
            }
        },
        "/auth/hello/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Return HelloWorld as sample API call",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "Return HelloWorld",
                "responses": {}
            }
        },
...
        "/items/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "List Items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Item"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Create Item",
                "parameters": [
                    {
                        "description": "Item Representation",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Item"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/items/backoffice": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "List Items for backoffice overview",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "No donation and transaction cost items",
                        "name": "skipHiddenItems",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "No license items",
                        "name": "skipLicenses",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Item"
                            }
                        }
                    }
                }
            }
        },
        "/items/{id}/": {
            "put": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Requires multipart form (for image)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Update Item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item Representation",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Item"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Delete Item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/mail-templates/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MailTemplates"
                ],
                "summary": "List mail templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.MailTemplate"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MailTemplates"
                ],
                "summary": "Create or update a mail template",
                "parameters": [
                    {
                        "description": "Template",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.MailTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mail-templates/{name}/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MailTemplates"
                ],
                "summary": "Get mail template by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.MailTemplate"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "tags": [
                    "MailTemplates"
                ],
                "summary": "Delete a mail template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mail-templates/{name}/send/": {
            "post": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MailTemplates"
                ],
                "summary": "Send a test email for a stored template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload with 'to' (array of recipients) and 'data' for template rendering",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/map/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Get longitudes and latitudes of all vendors for online map",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Map"
                ],
                "summary": "Get longitudes and latitudes of all vendors for online map",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.LocationData"
                            }
                        }
                    }
                }
            }
        },
        "/orders/": {
            "post": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Submits payment order to provider \u0026 saves it to database. Entries need to have an item id and a quantity (for entries without a price like tips, the quantity is the amount of cents). If no user is given, the order is anonymous.\nList all unverified orders",
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Orders",
                    "Orders"
                ],
                "summary": "List unverified orders",
                "parameters": [
                    {
                        "description": "Payment Order",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.createOrderResponse"
                        }
                    }
                }
            }
        },
        "/orders/unverified/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Submits payment order to provider \u0026 saves it to database. Entries need to have an item id and a quantity (for entries without a price like tips, the quantity is the amount of cents). If no user is given, the order is anonymous.\nList all unverified orders",
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Orders",
                    "Orders"
                ],
                "summary": "List unverified orders",
                "parameters": [
                    {
                        "description": "Payment Order",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.createOrderResponse"
                        }
                    }
                }
            }
        },
        "/orders/unverified/code/{orderCode}/transactionID": {
            "post": {
                "security": [
                    {
                        "KeycloakAuth": []
                    },
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Manually adds a transaction ID to an unverified order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Add Transaction ID to Order (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Code",
                        "name": "orderCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.addTransactionIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/orders/unverified/code/{orderCode}/verify/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    },
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Verifies order and creates payments manually",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Verify Payment Order by Code (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Code",
                        "name": "orderCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.VerifyPaymentOrderResponse"
                        }
                    }
                }
            }
        },
        "/orders/verify/": {
            "get": {
                "description": "Verifies order and creates payments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Verify Payment Order",
                "parameters": [
                    {
                        "type": "string",
                        "format": "3043685539722561",
                        "description": "Order Code",
                        "name": "s",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "882d641c-01cc-442f-b894-2b51250340b5",
                        "description": "Transaction ID",
                        "name": "t",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.VerifyPaymentOrderResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/forpayout/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    },
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Payments that do not have an associated payout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get list of all payments for payout",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2006-01-02T15:04:05Z",
                        "description": "Minimum date (RFC3339, UTC)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2006-01-02T15:04:05Z",
                        "description": "Maximum date (RFC3339, UTC)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Vendor LicenseID",
                        "name": "vendor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.paymentsResponse"
                        }
                    }
                }
            }
        },
...
        "/vendors/check/{licenseID}/": {
            "get": {
                "description": "Check if license id exists, return first name of vendor if it does",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "Check for license id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "License ID",
                        "name": "licenseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.addTransactionIDRequest": {
            "type": "object",
            "properties": {
                "transactionID": {
                    "type": "string"
                }
            }
        },
...
        "handlers.createOrderResponse": {
            "type": "object",
            "properties": {
                "smartCheckoutURL": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "KeycloakAuth": {
            "description": "\u003cb\u003ehow to generate an api key\u003c/b\u003e \u003cbr/\u003e\u003cbr/\u003e\u003ccode\u003ecurl -d 'client_id=frontend' -d 'scope=openid' -d 'username=test_superuser' -d 'password=Test123!' -d 'grant_type=password' 'http://keycloak:8080/realms/augustin/protocol/openid-connect/token' |     python3 -m json.tool | grep access_token\u003c/code\u003e\u003cbr/\u003e\u003cbr/\u003e\u003cbr/\u003eInsert the output into the field below the value of the access_token field.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}